move otelrestful sdk test
